Add ShareToken to funds
This commit aims to add in a converted version of the MelonPort
share contract to our funds. The MelonPort contract relies on a
few libraries and patterns that differ from what we are using
(main issue is that they do not appear to use zeppelin). The major
differences between our ShareToken.sol and their Shares.sol are
firstly the change in math libraries (they use DSMath, we use
SafeMath), the custom minting/burning functions created (I have
therefore made use of the MintableToken/BurnableToken libraries
in zeppelin) and the return type of a few functions (they were
very strong on using byteX types, wheras we have been making use
of string types).

There has also been a comment added to the FundFunctional contract
to clarify who actually "owns" the contract with the current build
flow, to ensure that it is more immediately obvious to future users.